switch to int

+++ Assets/Scripts/Scriptable Objects/Remote Data/GameSettingsScriptableObject.cs
@@ -25,7 +25,7 @@ namespace StarSalvager.ScriptableObjects
         public float botHorizontalSpeed = 30.0f;
         public float missionReminderFrequency = 25.0f;
         public float bonusShapeSpeed = 12.0f;
-        public float bonusShapeGearsReward = 100.0f;
+        public int bonusShapeGearsReward = 100;
 
         public bool cameraUseInputMotion = true;
         [ShowIf("cameraUseInputMotion", true)]
